wip: migration for task member time_spent

+++ bluebottle/tasks/models.py
@@ -314,9 +314,9 @@ class TaskMember(models.Model, PreviousStatusMixin):
                               default=TaskMemberStatuses.applied)
     motivation = models.TextField(_('Motivation'), help_text=_('Motivation by applicant.'), blank=True)
     comment = models.TextField(_('Comment'), help_text=_('Comment by task owner.'), blank=True)
-    time_spent = models.PositiveSmallIntegerField(_('time spent'),
-                                                  default=0,
-                                                  help_text=_('Time spent executing this task.'))
+    time_spent = models.FloatField(_('time spent'),
+                                     default=0,
+                                     help_text=_('Time spent executing this task.'))
 
     externals = models.PositiveSmallIntegerField(_('Externals'),
                                                  default=0,
